refactor(storage): move command owner module
